+++ DotDoc.EntityFrameworkCore.Extensions/ExecuteUpdate/ExecuteUpdateInterceptor.cs
@@ -2,7 +2,6 @@
 // This file is licensed to you under the MIT license.
 // See the License.txt file in the solution root for more information.
 
-using DotDoc.EntityFrameworkCore.Extensions.CustomQueryGenerators;
 using Microsoft.EntityFrameworkCore.Diagnostics;
 using System.Data.Common;